+++ client/src/graphql/queries/question.js
@@ -1,0 +1,18 @@
+import {gql} from '@apollo/client'
+
+export const GET_QUESTIONS_QUERY = gql`
+	query($formId: ID!) {
+		getQuestions(formId: $formId) {
+			id
+			title
+			description
+			type
+			answers {
+				id
+				body
+				correct
+			}
+			required
+		}
+	}
+`